// prisma/src/zod/schemas/objects/LocalUserWhereInput.schema.ts
import * as z from 'zod';
import type { Prisma } from '@prisma/client';
import { IntFilterObjectSchema as IntFilterObjectSchema } from './IntFilter.schema';
import { StringNullableFilterObjectSchema as StringNullableFilterObjectSchema } from './StringNullableFilter.schema';
import { StringFilterObjectSchema as StringFilterObjectSchema } from './StringFilter.schema';
import { DateTimeFilterObjectSchema as DateTimeFilterObjectSchema } from './DateTimeFilter.schema'

const localuserwhereinputSchema = z.object({
  AND: z.union([z.lazy(() => LocalUserWhereInputObjectSchema), z.lazy(() => LocalUserWhereInputObjectSchema).array()]).optional(),
  OR: z.lazy(() => LocalUserWhereInputObjectSchema).array().optional(),
  NOT: z.union([z.lazy(() => LocalUserWhereInputObjectSchema), z.lazy(() => LocalUserWhereInputObjectSchema).array()]).optional(),
  id: z.union([z.lazy(() => IntFilterObjectSchema), z.number().int()]).optional(),
  name: z.union([z.lazy(() => StringNullableFilterObjectSchema), z.string().max(255)]).optional().nullable(),
  email: z.union([z.lazy(() => StringFilterObjectSchema), z.string().max(255)]).optional(),
  passwordHash: z.union([z.lazy(() => StringFilterObjectSchema), z.string()]).optional(),
  role: z.union([z.lazy(() => StringFilterObjectSchema), z.string().max(50)]).optional(),
  permissions: z.union([z.lazy(() => StringNullableFilterObjectSchema), z.string().max(255)]).optional().nullable(),
  createdAt: z.union([z.lazy(() => DateTimeFilterObjectSchema), z.coerce.date()]).optional(),
  updatedAt: z.union([z.lazy(() => DateTimeFilterObjectSchema), z.coerce.date()]).optional(),
  locale: z.union([z.lazy(() => StringFilterObjectSchema), z.string()]).optional()
}).strict();
export const LocalUserWhereInputObjectSchema: z.ZodType<Prisma.LocalUserWhereInput> = localuserwhereinputSchema as unknown as z.ZodType<Prisma.LocalUserWhereInput>;
export const LocalUserWhereInputObjectZodSchema = localuserwhereinputSchema;